+++ examples/fuel_map/simplecase.py
@@ -15,10 +15,10 @@ my_fuelmap = FuelMap(fuel_db=my_db)
 
 # add a fuel patch of the tall grass fuel from the FireFluxI database
 ## to show the available fuel keys of the database, use print(my_db)
-my_fuelmap.add_fuel_rectangle_patch(xpos=[50, 450], ypos=[50, 450], fuel_key="FireFluxI_tall_grass")
+my_fuelmap.add_fuel_rectangle_patch(pos1=[50, 450], pos2=[50, 450], fuel_key="FireFluxI_tall_grass")
 
 # add a ignition patch
-my_fuelmap.add_ignition_rectangle_patch(xpos=[100, 105], ypos=[245, 255], ignition_time=10)
+my_fuelmap.add_ignition_rectangle_patch(pos1=[100, 105], pos2=[245, 255], ignition_time=10)
 
 # dump for mesonh
 my_fuelmap.dump_mesonh()